Subindo ajustes.
