+++ PORTIFOLIO/PORTIFOLIO.API/CONTROLLER/ChatManagerController.cs
@@ -1,8 +1,8 @@
-﻿using APPLICATION.DOMAIN.CONTRACTS.SERVICES.CHAT;
+﻿using APPLICATION.APPLICATION.SIGNALR;
+using APPLICATION.DOMAIN.CONTRACTS.SERVICES.CHAT;
 using APPLICATION.DOMAIN.DTOS.CHAT;
 using APPLICATION.DOMAIN.DTOS.RESPONSE.UTILS;
 using APPLICATION.DOMAIN.UTILS.EXTENSIONS;
-using APPLICATION.INFRAESTRUTURE.SIGNALR;
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.SignalR;
