+++ PORTIFOLIO/PORTIFOLIO.API/CONTROLLER/UserManagerController.cs
@@ -1,4 +1,5 @@
-﻿using APPLICATION.DOMAIN.CONTRACTS.SERVICES.USER;
+﻿using APPLICATION.APPLICATION.SIGNALR;
+using APPLICATION.DOMAIN.CONTRACTS.SERVICES.USER;
 using APPLICATION.DOMAIN.DTOS.CONFIGURATION.AUTH.CUSTOMAUTHORIZE.ATTRIBUTE;
 using APPLICATION.DOMAIN.DTOS.CONFIGURATION.AUTH.TOKEN;
 using APPLICATION.DOMAIN.DTOS.REQUEST.USER;
@@ -8,7 +9,6 @@ using APPLICATION.DOMAIN.ENUMS;
 using APPLICATION.DOMAIN.HUB;
 using APPLICATION.DOMAIN.UTILS.EXTENSIONS;
 using APPLICATION.DOMAIN.UTILS.GLOBAL;
-using APPLICATION.INFRAESTRUTURE.SIGNALR;
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.SignalR;
